Fixed intValue method: should return int, not double
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@3818 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/qsar/result/IntegerArrayResult.java
@@ -48,7 +48,7 @@ public class IntegerArrayResult implements DescriptorResult {
     /**
      * The first int is at index = 0;
      */
-    public double get(int index) {
+    public int get(int index) {
         return ((Integer)this.array.get(index)).intValue();
     }
